frontend/requirements.txt
joblib~=1.2.0
matplotlib~=3.7.0
pandas~=1.5.3
numpy~=1.23.5
optuna~=3.1.0
plotly~=5.13.1
PyYAML~=6.0
requests~=2.28.2
seaborn~=0.12.2
streamlit~=1.21.0
scikit-learn~=1.2.1